separated out calculator function for premium and stored it in config folder
